More cleanup. Removed majority of unnecessary casts.

+++ src/API/com/bioxx/tfc/api/Crafting/BarrelBriningRecipe.java
@@ -19,7 +19,7 @@ public class BarrelBriningRecipe extends BarrelRecipe
 	{
 		if(item != null && item.getItem() instanceof IFood && !Food.isBrined(item))
 		{
-			float w = (float)((IFood)item.getItem()).getFoodWeight(item);
+			float w = ((IFood) item.getItem()).getFoodWeight(item);
 			if(fluid.isFluidEqual(recipeFluid) && (w <= 1f*(fluid.amount/this.recipeFluid.amount)))
 			{
 				return true;
