+++ src/Common/com/bioxx/tfc/Blocks/Terrain/BlockCobble.java
@@ -154,7 +154,7 @@ public class BlockCobble extends BlockTerra
 
 			if(!canFallOneBelow && (count > 2) && sides.size() >= 1)
 			{
-				switch((Integer)sides.get(random.nextInt(sides.size())))
+				switch (sides.get(random.nextInt(sides.size())))
 				{
 				case 0:
 				{
