+++ src/Common/com/bioxx/tfc/Blocks/Vanilla/BlockCustomFenceGate.java
@@ -48,7 +48,7 @@ public class BlockCustomFenceGate extends BlockFenceGate implements ITileEntityP
 
 		int l = (te != null && te instanceof TEFenceGate /*Fixes ClassCastException*/) ? (((TEFenceGate) (te)).getDirection()) : 0;
 		boolean open = (te != null) ? (((TEFenceGate) (te)).getOpen()) : false;
-		return open ? null : (l != 2 && l != 0 ? AxisAlignedBB.getBoundingBox((double) ((float) x + 0.375F), (double) y, (double) z, (double) ((float) x + 0.625F), (double) ((float) y + 1.5F), (double) (z + 1)) : AxisAlignedBB.getBoundingBox((double) x, (double) y, (double) ((float) z + 0.375F), (double) (x + 1), (double) ((float) y + 1.5F), (double) ((float) z + 0.625F)));
+		return open ? null : (l != 2 && l != 0 ? AxisAlignedBB.getBoundingBox(x + 0.375F, y, z, x + 0.625F, y + 1.5F, z + 1) : AxisAlignedBB.getBoundingBox(x, y, z + 0.375F, x + 1, y + 1.5F, z + 0.625F));
 	}
 
 	@SideOnly(Side.CLIENT)
@@ -114,7 +114,7 @@ public class BlockCustomFenceGate extends BlockFenceGate implements ITileEntityP
 	@Override
 	public void onBlockPlacedBy(World par1World, int par2, int par3, int par4, EntityLivingBase par5EntityLivingBase, ItemStack par6ItemStack)
 	{
-		int l = (MathHelper.floor_double((double)(par5EntityLivingBase.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3) % 4;
+		int l = (MathHelper.floor_double(par5EntityLivingBase.rotationYaw * 4.0F / 360.0F + 0.5D) & 3) % 4;
 		((TEFenceGate)(par1World.getTileEntity(par2, par3, par4))).setDirection((byte)l);
 
 		//par1World.setBlockMetadataWithNotify(par2, par3, par4, l, 2);
